+++ app/src/main/java/net/gas/gascontact/business/database/entities/Units.kt
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "units")
 data class Units (
-    @PrimaryKey val id: Short,
+    @PrimaryKey val id: Int,
     var name: String?,
     val code: String?,
     val rangir: Int?
